Fixed prediction_plotter for sequence trainers, removed old trainers.
